script/lib/distro.rb: reorder alphanumerically
We maintain a compatibility matrix of Linux distributions and versions
in the DistroMap Ruby class in our script/lib/distro.rb source file,
which is utilized by several scripts run by our GitHub Actions release
workflow jobs when building and naming the Linux packages we publish
on Packagecloud.

We will update the matrix of Linux distribution versions we support
in a subsequent commit in this PR.  Prior to making these changes,
though, we first reorder the list of distribution releases that we
consider to be compatible with the package we build for Debian 11.
Sorting this list alphanumerically ensures that it aligns with the
other lists defined in our DistroMap class.