script/lib/distro.rb: move openSUSE 15.6 to RHEL 9
We maintain a compatibility matrix of Linux distributions and versions
in the DistroMap Ruby class in our script/lib/distro.rb source file,
which is utilized by several scripts run by our GitHub Actions release
workflow jobs when building and naming the Linux packages we publish
on Packagecloud.

In commit 5ed62ba5e9bdf7cc370a516ae320a6fd4ff0a60e in PR #5647 we
revised this matrix so as to reassign openSUSE Leap 15.5 and 15.6 as
well as SUSE Linux Enterprise Server (SLES) 15 SP5 to the set of
distribution versions that we consider to be compatible with the RPM
packages we build for Red Hat Enterprise Linux (RHEL) 8, CentOS 8,
and Rocky Linux 8.

This works because the version of the GNU C library (glibc) provided by
the RHEL 8 family of distributions is 2.28, while the version of glibc
available on the more recent SUSE distributions is at least 2.31.  The
GNU C library keeps its ABI backwards-compatible, so any Git LFS binary
we build against glibc v2.28 on RHEL/CentOS 8 should also be functional
on the SUSE releases with glibc v2.31 or higher.

In a subsequent commit in this PR we will update the matrix in our
DistroMap Ruby class to remove Linux distribution versions that have
reached the end of their regular support lifecycle and to add new
versions of several distributions, including SLES 15 SP6.

Prior to making these changes, though, we first reassign OpenSUSE
Leap 15.6 to the set of distribution versions that we consider to be
compatible with the RPM packages we build for RHEL 9 and Rocky Linux 9,
as we would like to also include SLES 15 SP6 in the same set.

Both of these recent SUSE distributions versions provide glibc version
2.38, which is newer than the glibc version 2.34 included in the RHEL 9
family of distributions.  Thus a Git LFS binary we build against the
ABI of glibc v2.34 on RHEL/Rocky Linux 9 should be functional on the
latest SUSE releases as well.

We have confirmed that this is the case by successfully running the
full Git LFS test suite on both openSUSE Leap 15.6 and SLES 15 SP6
using a Git LFS v3.5.1 binary installed from the RPM we built RHEL 9
and Rocky Linux 9.

Based on these results, we can move openSUSE Leap 15.6 into the
set of distribution versions for which the RPM package we build
for RHEL 9 and Rocky Linux 9 will be compatible.