script/lib/distro.rb: reformat EOL notes
We maintain a compatibility matrix of Linux distributions and versions
in the DistroMap Ruby class in our script/lib/distro.rb source file,
which is utilized by several scripts run by our GitHub Actions release
workflow jobs when building and naming the Linux packages we publish
on Packagecloud.

We will update the matrix of Linux distribution versions we support
in a subsequent commit in this PR.  Prior to making those changes,
we first reformat the code comments we use to track the end of
each distribution version's regular support lifecycle so that all
the comments align with each other.  This will make comparisons and
updates slightly easier, particularly when we move an entry in our
matrix from one list of entries to another.