script/lib/distro.rb: update distribution list
We maintain a compatibility matrix of Linux distributions and versions
in the DistroMap Ruby class in our script/lib/distro.rb source file,
which is utilized by several scripts run by our GitHub Actions release
workflow jobs when building and naming the Linux packages we publish
on Packagecloud.

As we anticipate making a v3.6.0 release of the Git LFS client in
the near future, we first update the list of Linux distribution
versions for which we will build RPM and Debian packages, removing
those which have reached the end of their regular support lifecycles
and adding those which have become available since our v3.5.0 release.

As noted in a previous commit in this PR, we are adding SUSE Linux
Enterprise Server (SLES) 15 SP6 to the set of distribution versions
that we consider to be compatible with the RPM packages we build
for Red Hat Enterprise Linux (RHEL) 9 and Rocky Linux 9.  This is
because the version of the GNU C library (glibc) provided by the
RHEL 9 family of distributions is 2.34, while the version of glibc
available on SLES 15 SP6 is 2.38, and any Git LFS binary we build
for the former will be functional with the latter as well.

While most of our updates are unexceptional, there are two notable
distribution versions that have reached the end of their regular support
lifecycles but which we leave in our matrix for the time being.

First, we leave the Debian 10 ("buster") release in our matrix
because it is the base for a number of other distribution versions
that have not yet reached the end of their regular support lifecycles.
In particular, Ubuntu 20.04 LTS (Focal Fossa) and the Linux Mint 20.x
series of releases will only reach the end of their regular support
periods in April 2025.  Since our scripts generate Git LFS release
notes and links to the Packagecloud service which depend on the
name of the base component of each set of related distribution versions,
if we removed the "debian/buster" entry we would also have to rename
various other fields and or generate a "Debian 10" link in our release
notes which did not refer to a "debian/*" package.  Hence we simply
leave the "debian/buster" entry in place until after April 2025, at
which point we will remove all the distribution versions based on
Debian 10.

Second, we do not remove the entire set of distribution versions
similar to RHEL 7 because the last one to reach the end of its regular
support lifecycle, SLES 12 SP5, only did so very recently at the end
of last month.  Further, other than a brief mention in PR #5647,
we have not announced that we would no longer build any packages for
this family of distributions.  For this reason, we will leave the
SLES 12 SP5 entry and the base RHEL 7 entry in place for one final
Git LFS minor release cycle, after which they will be removed.

To clarify that further, we revise our comment tracking the support
lifecycle of SLES 12 SP5, which has previously included both the end
of the regular support period and the end of SUSE's Long Term Service
Pack Support (LTSS) for the release.  However, our practice for all
other Linux distributions has been to only provide packages during the
regular support period, and not during any form of extended support as
may be offered by some commercial vendors.  For instance, we stopped
providing packages for Ubuntu 18.04 LTS (Bionic Beaver) after the end
of its normal long-term support in May 2023, although ESM (Expanded
Security Maintenance) support is still available through April 2028.
We therefore we now remove any suggestion that we will continue to
support SLES 12 SP5 throughout its extended LTSS period.