t: avoid incorrect negated test expressions
In commit a5d20de4a8de9660720c79858ae57e40cffa1def of PR #5282 we
revised a number of our test scripts so as to avoid the use of the
"!" shell word in cases where we want to confirm that a command fails
and returns a non-zero exit code.  We made this change because we had
a variety of instances where we had checks of this form which would
in practice always pass, regardless of the exit code of their commands.

(This behaviour is a consequence of our use of "set -e" in all our
test scripts.  The Git project's shell test scripts do not use this
option, but ours do; it means that the shell will exit immediately
if a normal command returns a non-zero exit code.  However, according
to the POSIX shell specification, when the "set -e" option is enabled
and a command's exit code is inverted with "!", the shell will not exit
immediately even if the inverted value is non-zero.)

Commit a5d20de4a8de9660720c79858ae57e40cffa1def addressed all of the
instances in our test scripts where we relied on an incorrect usage
of the "!" shell word.  However, we have another similar form of
incorrect usage in a number of our test scripts where we rely on the
"!" test operator to confirm that a command's exit code is non-zero
(i.e., in conditions where we expect the command to fail).

In particular, we often use tests of the following form, with the
expectation that "foo" should not appear in the log file, and if it
does the test should fail:

  [ ! $(grep "foo" bar.log) ]

These checks always succeed, but for the wrong reason.  When the
grep(1) command does not find the given string in its input stream
or the specified file, it prints nothing to its standard output and
returns a non-zero exit code.  Because we are evaluating a command
substitution inside the test expression, and the output of that
substitution is the standard output of the command, which is empty,
the test devolves to an expression of the form:

  [ ! ]

The Bash shell treats the absence of an expression as a false value,
which is then inverted by the "!" operator, so the test passes, but
not because has confirmed that the command's exit code was non-zero.

Some of these improper test expressions could still catch regressions
in the Git LFS client's behaviour, though.  For instance, in the
example above, if "foo" was found in the log file, the test expression
that would be evaluated would have the following form, where all the
log lines which matched the string "foo" would appear after the "!"
operator:

  [ ! ... foo ... ]

This might form an invalid shell expression, or one with too many
arguments (thus causing a "[: too many arguments" error), in which
case the test would fail.

We can replace all the instances where we rely on these improper
test expressions using the same approach we used in commit
a5d20de4a8de9660720c79858ae57e40cffa1def.  We simply perform the
commands from what was enclosed in a command substitution as the first
command pipeline in an AND list, where "exit 1" is the second and
final command in the list.

If the first command pipeline returns a non-zero exit code, as we
expect it to, the "exit 1" is not executed.  However, if the pipeline
should catch a regression and unexpectedly succeed, "exit 1" will be
executed, causing our test to fail.  Note that we do not rely on the
"set -e" option here, because it ignores the return values of command
lists entirely.

In cases where our revised check is the last line of one of our tests,
we add a "true" statement, because otherwise the test will fail since
the last evaluated command would be the first one in the list, i.e.,
the grep(1) command that we expect will find no matches in its input.

In our t/t-migrate-import.sh test script we also have some improper
test expresssions of the following form, where we expect that the
output of the "git cat-file" command should be empty because the
".gitattributes" file at the given Git ref is blank:

  [ ! $(git cat-file -p "ref:.gitattributes") ]

As with the improper tests of the output of grep(1) commands, these
tests succeed but for the wrong reason, because they also devolve
to expressions of the form:

  [ ! ]

To repair these checks, we simply replace the "!" test operator
with the "-z" one, as we just want to confirm that the output
of the command substitution is a zero-length string.