t/t-clone.sh: clarify test of encrypted TLS key
We added support for encrypted TLS/SSL certificate keys in PR #3270, and
as part of that change, in commit 706becaf20fc7657dcd296a17234c875f546603d
we updated the "clone ClientCert" test in our t/t-clone.sh test script to
perform "git lfs clone" operations twice, once with an unencrypted TLS/SSL
certificate key and once with an encrypted one.

Our test needs to ensure that Git and Git LFS do not generate user
prompts to retrieve the passphrase to the encrypted key, as this would
make impossible to run the test in our CI jobs, and so we set the
GIT_SSL_CERT_PASSWORD_PROTECTED environment variable to "1" at the
start of the main loop in our test.

However, we do this for both passes through the loop, which means we
set the GIT_SSL_CERT_PASSWORD_PROTECTED environment variable when
testing with the unencrypted TLS/SSL certificate key as well as with
the encrypted one.  We only need to set it in the latter case, though.

To clarify the intent of the test's loops, we refactor the test to now
iterate through a list of two values, "false" and "true".  When the
value is "true", we set the "http.sslCertPasswordProtected" Git
configuration option to "true", and reset the "http.<url>.sslKey" option
with the path from the LFS_CLIENT_KEY_FILE_ENCRYPTED variable.

By refactoring the test's main loop in this way we can avoid setting
Git environment variables or configuration options which do not apply
when neither the certificate file nor the private key file are encrypted.

In a prior commit this PR we moved the "git config" commands which set
the "http.<url>.sslCert" and "http.<url>.sslKey" configuration options
specific to the "clone ClientCert" test out of the common setup() function
in our t/testhelpers.sh test library and into the test itself.

To align with those initial "git config" commands and with the command
we use on the second pass through the test's loop to reset the value
of the "http.<url>.sslKey" option, we prefer to set the
"http.sslCertPasswordProtected" Git configuration option instead of the
GIT_SSL_CERT_PASSWORD_PROTECTED environment variable.  This option has
the same effect on Git's behaviour as the environment variable, so there
is no change to the test conditions.

We also take the opportunity to fix a small typo in the directory
name argument passed to the "git lfs clone" command, and we avoid
the need to delete the working tree created by the previous iteration
of the loop by including our loop index value in the directory name.

Finally, we add a detailed set of code comments to explain exactly
why we have to create two credential record files for use by our
git-credential-lfstest helper program, associated with the paths of
both the TLS/SSL certificate file and the private key file, when
only the latter is encrypted.  This is due to the fact that Git and
Git LFS each make calls to the credential helper to retrieve the
private key's passphrase, but do so for different reasons, and use
different "path" values in the input they send to the "git credential
fill" command.