t/t-filter-process.sh: update Git version comment
Because full support of the "filter" attribute and protocol were only
introduced with Git version 2.11.0, we require at least that version of
Git for our t/t-filter-process.sh test script.  This attribute and
the associated protocol were developed in part so as to make the
Git LFS client's integration with Git more efficient.

The original version of the t/t-filter-process.sh test script predated
the release of Git v2.11.0, so when it was introduced in commit
d1dca3ee2deb28b9e24ffb6561c8d21af1b433d1 of PR #1617, it included
a workaround technique for detecting whether the version of Git in use
did or did not support the "filter" attribute yet.  This workaround
was preceded by a comment describing how it worked.

In commit 54ddc13bffed2b0303e8818bb8cbde4bea469583 of PR #1699 the
workaround was removed and replaced with a call to the
ensure_git_version_isnt() function from our t/testhelpers.sh shell
library, since Git version 2.11.0 had been released and so we could
perform a simpler version check to confirm whether Git supported the
"filter" attribute.

However, the comment describing the original workaround technique
for detecting support for the "filter" attribute was left in place,
so we update it now.