t/t-clone.sh: expand checks of Git clone commands
In previous commits in this PR we have refactored a number of the
initial tests in our t/t-clone.sh script so they are more consistent
with each other and perform a more complete set of checks of the
"git lfs clone" command when it is used with HTTP and HTTPS remote
URLs and with TLS/SSL client certificates.

All four of these tests run one or more "git lfs clone" commands.  After
each, they check the logs from the command, and then verify that a local
Git repository has been created with the expected Git LFS hooks installed
and with files in the working tree populated from the contents of Git LFS
objects that have been downloaded.

Three of these four tests, other than the very first "clone" test, also
follow a pattern added in commits 4c64e8212454f248478f2d431f3e6931e3b67fe6
and 8f91a1bc7fe0c74cde254b432302e6ac871d1a86 of PR #1067, in which they
run a "git clone" command after the "git lfs clone" command.  This test
design was introduced first for the "cloneSSL" test, and then replicated
in the "clone ClientCert" test when that was added in commit
daba49ae43340ee30b2e7c9a6b8d5927dbd71f51 of PR #1893.  In prior commits
in this PR we have now replicated it again into the "clone ClientCert
with homedir certs" test, and have also ensured that it runs twice in
both that test and the "clone ClientCert" test, once with an unencrypted
TLS/SSL private key file and once with an encrypted one.

However, although these three tests run a "git clone" command, they
simply check that it does not exit with a non-zero error code (because
we use the "set -e" option in our tests, so if the command failed it
would cause the tests to fail).  This is less than ideal, given that the
"git lfs clone" command is now deprecated, and almost all Git LFS users
will use the regular "git clone" command instead.

As suggested by larsxschneider during the review of this PR, we can
improve all four of these tests by adding checks following each "git clone"
command similar to those we perform after the "git lfs clone" commands.

We can also add a "git clone" command to the end of the initial "clone"
test so that it now follows the same pattern as the other three tests.

In each of these tests, we now confirm that the "git clone" command
exits with a non-zero value, but also that it outputs a "Cloning into"
message.  We then check that Git LFS hooks were installed in the cloned
local repository, that the expected number of Git LFS objects were
downloaded, and that files in the working tree have been populated
with the contents of those objects.