docs/howto/release-git-lfs.md: sort Linux assets
In commit d61b3ead04be8421fa149507000ebfcb48cec2ae of PR #5607 we added
support in our Makefile and scripts to build and upload a version of our
Linux release assets for the LoongArch64 platform, and at the same time
we added an entry for this platform to the end of the list of Linux assets
in our release process documentation.

Before updating the documentation of our release process, we first reorder
this list to keep it in alphanumeric order, so as to match how assets are
listed in our public release notes.