docs/howto/release-git-lfs.md: add RISC-V 64 asset
In PR #5438 we added support in our Makefile and scripts to build and
upload a version of our Linux release assets for the RISC-V 64 platform.

As the documentation of our release process includes a list of the
Linux assets we build, we now add an entry for the RISC-V 64 platform
so as to match the assets listed in our public release notes.