docs/howto/release-git-lfs.md: update some steps
Based on the recent experience of releasing the 3.6.0 version of Git LFS,
we update several steps in our documentation to more accurately reflect
our current release process.

First, we add a reminder that we should include in our release notes
any changes from the previous Git LFS release in terms of platform
requirements or the set of operating system versions we support.

As well, we more fully document the dependencies of our script/finalize
script, and note that a PR to update Homebrew should now be generated
automatically within a few hours of our making a new release public.