docs/howto/release-git-lfs.md: add Go upgrade step
Before we release a new version of Git LFS, we generally want to upgrade
the version of Go used in the Docker container images generated by
our git-lfs/build-dockers repository, so we add notes regarding this
step to our release process documentation.

Although Markdown does not require us to renumber the items in an
ordered list, we do so anyway as this makes our documentation more
readable when viewed as a plain text file.