script/upload: use generic sha3sum input format
The verify_assets() function in our script/upload script performs a
final verification of the contents of the GPG-signed "sha25sum.asc" and
"hashes.asc" files we include in the assets attached to our public
release notes.

To perform these checks, it runs the shasum(1) utility for all the SHA-2
hashes, the sha3sum(1) utility for the SHA-3 hashes, and the b2sum(1)
utility for the BLAKE2b hashes, passing each utility the appropriate
lines from the "hashes.asc" file.  These lines have the format output
by our script/hash-files script, namely:

  <hash-name> (<file-name>) = <hash-value>

While this format is accepted and parsed successfully by all of the
utilities on most Linux systems, the sha3sum(1) program available on
macOS via Homebrew does not permit this input format with its -c option.
Instead, it requires as input exactly the same format as it outputs when
hashing a set of files, i.e.:

  <hash-value>  <file-name>

This disparity is due to the fact that Homebrew packages as the "sha3sum"
utility a different implementation than is packaged for Debian Linux.
Specifically, the Homebrew package is built from the following source
project:

  https://codeberg.org/maandree/sha3sum
  https://formulae.brew.sh/formula/sha3sum

On Debian, however, the sha3sum(1p) utility is a Perl script provided
with the Digest::SHA3 Perl extension:

  https://metacpan.org/dist/Digest-SHA3/view/sha3sum
  https://sources.debian.org/src/libdigest-sha3-perl/
  https://manpages.debian.org/unstable/libdigest-sha3-perl/sha3sum.1p.en.html

This Perl script accepts a wider range of input formats than the program
packaged by Homebrew, including the one created by our script/hash-files
script.  It is also able to deduce the SHA-3 algorithm to use based on
the hash name in each line, which the sha3sum(1) utility on macOS is
unable to do.

As we sometimes want to run our scripts on macOS when generating a new
Git LFS release, we adjust our script/upload script so it rewrites the
lines from the "hashes.asc" file into the format accepted by the Homebrew
version of sha3sum(1), and also so it invokes that utility separately
for each hash algorithm.  As this also works with the Perl script version
of sha3sum(1p), our script/upload script should now succeed on both
macOS and Linux in the future.